requirements.txt
Flask==2.0.1
python-dotenv==0.19.2
Werkzeug==2.0.3
azure-storage-blob==12.10.0